requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.3

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2

# Machine learning
scikit-learn>=1.2.2
nltk>=3.8.1

# Document parsing
pdfplumber>=0.9.0
pytesseract>=0.3.10
openpyxl>=3.1.2
tabula-py>=2.7.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
